An attempt to transcode and stream media via http ranges

+++ main.go
@@ -6,8 +6,12 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"os/exec"
+	"path"
+	"path/filepath"
 	"strconv"
 	"strings"
+	"time"
 
 	cli "github.com/urfave/cli"
 )
@@ -216,7 +220,7 @@ func main() {
 				for _, filename := range playlist {
 					contentType, _ := getLikelyContentType(filename)
 					fmt.Printf("Playing '%s'\n", filename)
-					err := castApplication.PlayMedia(folder+filename, contentType)
+					err := castApplication.PlayMedia(folder+filename, contentType, false)
 					fmt.Printf("error: %v\n", err)
 				}
 				return nil
@@ -231,60 +235,57 @@ func main() {
 			Action: func(c *cli.Context) error {
 				filenameOrUrl := c.Args().Get(0)
 				contentType := c.Args().Get(1)
+				trancodingVideo := false
 				fmt.Println(filenameOrUrl, contentType)
+
 				if contentType == "" {
 					var err error
 					contentType, err = getLikelyContentType(filenameOrUrl)
 					if err != nil {
 						log.Printf("Unable to find content type: %s", err)
-						return nil
-						/*
-							// https://github.com/pkg4go/httprange
-							// https://stackoverflow.com/questions/3303029/http-range-header
 
+						// For now only allow .avi files to be transcoded to .mp4
+						if ext := path.Ext(filename); ext != ".avi" {
+							log.Printf("Not able to transcode '%s' files to mp4\n", ext)
+							return nil
+						}
 
-							// From the stdout of the below command
-							//	Duration: 00:21:32.76
-							//
+						// Start transcoding ffmpeg to a file, and attempt to get the server to
+						// serve the file as it is transcoding
+						tmpDir, err := ioutil.TempDir("", "chromecast")
+						if err != nil {
+							log.Fatal(err)
+						}
+						defer os.RemoveAll(tmpDir)
 
-							// Start transcoding ffmpeg to a file, and attempt to get the server to
-							// serve the file as it is transcoding
+						writeFilepath := filepath.Join(tmpDir, fmt.Sprintf("%d-%d.mp4", filenameOrUrl, time.Now().Unix()))
+						cmd := exec.Command(
+							"ffmpeg",
+							"-i", filenameOrUrl,
+							"-vcodec", "h264",
+							"-f", "mp4",
+							"-movflags", "frag_keyframe+faststart",
+							"-strict", "-experimental",
+							writeFilepath,
+						)
 
-							tmpDir, err := ioutil.TempDir("", "chromecast")
-							if err != nil {
-								log.Fatal(err)
-							}
-							defer os.RemoveAll(tmpDir) // ??
-							writeFilepath := filepath.Join(tmpDir, fmt.Sprintf("test-%d.mp4", time.Now().Unix()))
+						fmt.Printf("Starting transcoding\n")
+						if err := cmd.Start(); err != nil {
+							log.Fatal(err)
+						}
+						fmt.Println("finished transcoding")
 
-							cmd := exec.Command(
-								"ffmpeg",
-								"-i", filenameOrUrl,
-								"-vcodec", "h264",
-								"-f", "mp4",
-								"-movflags", "frag_keyframe+faststart",
-								"-strict", "-experimental",
-								writeFilepath,
-								// "pipe:1", // This will pipe it to stdout
-							)
+						filenameOrUrl = writeFilepath
+						contentType = "video/mp4"
+						transcodingVideo = true
 
-							fmt.Printf("Starting transcoding\n")
-							if err := cmd.Run(); err != nil {
-								log.Fatal(err)
-							}
-							fmt.Println("finished transcoding")
+						// Give it some time to start transcoding the media
+						time.Sleep(time.Second * 10)
 
-							filenameOrUrl = writeFilepath
-							contentType = "video/mp4"
-
-							time.Sleep(time.Second * 10)
-						*/
-
-						//return err
 					}
 				}
 				fmt.Println(filenameOrUrl, contentType)
-				if err := castApplication.PlayMedia(filenameOrUrl, contentType); err != nil {
+				if err := castApplication.PlayMedia(filenameOrUrl, contentType, transcodingVideo); err != nil {
 					fmt.Printf("Error: %s\n", err)
 				}
 				return nil
